# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/data_consents:
    get:
      tags:
        - DataConsents
      summary: List DataConsents
      description: Get a paginated list of DataConsents with optional filtering and sorting
      operationId: listDataConsents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataConsentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DataConsents
      summary: Create DataConsent
      description: Create a new DataConsent
      operationId: createDataConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDataConsentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataConsent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/data_subject_requests:
    get:
      tags:
        - DataSubjectRequests
      summary: List DataSubjectRequests
      description: Get a paginated list of DataSubjectRequests with optional filtering and sorting
      operationId: listDataSubjectRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSubjectRequestList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DataSubjectRequests
      summary: Create DataSubjectRequest
      description: Create a new DataSubjectRequest
      operationId: createDataSubjectRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDataSubjectRequestRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSubjectRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    DataConsent:
      type: object
      required:
        - id
        - employee_id
        - data_category
        - lawful_basis
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        data_category:
          type: string
        lawful_basis:
          type: string
        consent_given_at:
          nullable: true
          type: string
          format: date-time
        consent_method:
          nullable: true
          type: string
        privacy_notice_version:
          nullable: true
          type: string
        retention_until:
          nullable: true
          type: string
          format: date
        withdrawn_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateDataConsentRequest:
      type: object
      required:
        - employee_id
        - data_category
        - lawful_basis
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        data_category:
          type: string
        lawful_basis:
          type: string
        consent_given_at:
          nullable: true
          type: string
          format: date-time
        consent_method:
          nullable: true
          type: string
        privacy_notice_version:
          nullable: true
          type: string
        retention_until:
          nullable: true
          type: string
          format: date
        withdrawn_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    DataConsentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DataConsent'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    DataSubjectRequest:
      type: object
      required:
        - id
        - employee_id
        - request_type
        - status
        - requested_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        request_type:
          type: string
        status:
          type: string
        requested_at:
          type: string
          format: date-time
        fulfilled_at:
          nullable: true
          type: string
          format: date-time
        response:
          nullable: true
          type: object
        note:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateDataSubjectRequestRequest:
      type: object
      required:
        - employee_id
        - request_type
        - status
        - requested_at
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        request_type:
          type: string
        status:
          type: string
        requested_at:
          type: string
          format: date-time
        fulfilled_at:
          nullable: true
          type: string
          format: date-time
        response:
          nullable: true
          type: object
        note:
          nullable: true
          type: string
        metadata:
          type: object
...
    DataSubjectRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DataSubjectRequest'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []